Fix postgres connection uri
The connection string `postgres://verifudge` looks to host verifudge to
expose a db on port 5432. This new connection string
`postgres:///verifudge` looks to localhost to expose a db named
verifudge on port 5432.

The difference is because the `///` causes the connector to assume
`//localhost/` so our connection string gets expanded to
`postgres://localhost/verifudge`.